backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
opencv-python-headless==4.9.0.80
ultralytics==8.1.0
numpy==1.24.3
Pillow==10.2.0
gunicorn==21.2.0